Time: 6 ms (36.72%), Space: 13.8 MB (12.3%) - LeetHub

+++ 0330-patching-array/0330-patching-array.cpp
@@ -2,8 +2,8 @@ class Solution {
 public:
     int minPatches(vector<int>& nums, int n) {
        long long miss = 1;
-        int result = 0;
-        size_t i = 0;
+        long result = 0;
+        long i = 0;
 
         while (miss <= n) {
             if (i < nums.size() && nums[i] <= miss) {
